wip: fixing ginkgo setup
