+++ test/snapshot/edge_classic_e2e/edge_classic_suite_test.go
@@ -1,0 +1,28 @@
+package edge_classic_e2e
+
+import (
+	"context"
+	"testing"
+
+	. "github.com/onsi/ginkgo/v2"
+	"github.com/solo-io/gloo/test/helpers"
+	skhelpers "github.com/solo-io/solo-kit/test/helpers"
+)
+
+func TestEdgeGateway(t *testing.T) {
+	helpers.RegisterGlooDebugLogPrintHandlerAndClearLogs()
+	skhelpers.RegisterCommonFailHandlers()
+	skhelpers.SetupLog()
+	RunSpecs(t, "Gloo Gateway Suite")
+}
+
+var (
+	ctx       context.Context
+	ctxCancel context.CancelFunc
+)
+
+var _ = BeforeSuite(StartTestHelper)
+
+func StartTestHelper() {
+	ctx, ctxCancel = context.WithCancel(context.Background())
+}
