--- test/snapshot/gloo_gateway_e2e/testcase/ingress2.go
@@ -1,15 +1,0 @@
-package testcase
-
-import (
-	. "github.com/onsi/ginkgo/v2"
-)
-
-var Test = func(input TestInput) {
-	BeforeEach(func() {
-		input.ApplyConfig()
-	})
-
-	It("Get 200 response", func() {
-
-	})
-}
