--- test/snapshot/gloo_gateway_int/outputs/http-routing-proxy.yaml
@@ -1,55 +1,0 @@
----
-listeners:
-  - aggregateListener:
-      httpFilterChains:
-        - matcher: {}
-          virtualHostRefs:
-            - http~httpbin.example.com
-            - http~example.com
-      httpResources:
-        virtualHosts:
-          http~reviews.example.com:
-            domains:
-              - httpbin.example.com
-            name: http~reviews.example.com
-            routes:
-              - matchers:
-                  - headers:
-                      - name: env
-                        value: canary
-                    prefix: /
-                options: {}
-                routeAction:
-                  single:
-                    upstream:
-                      name: default-httpbin-v1-svc-canary-8000
-                      namespace: default
-              - matchers:
-                  - prefix: /
-                options: {}
-                routeAction:
-                  single:
-                    upstream:
-                      name: default-httpbin-svc-8000
-                      namespace: default
-          http~example.com:
-            domains:
-              - example.com
-            name: http~example.com
-            routes:
-              - matchers:
-                  - prefix: /
-                options: {}
-                routeAction:
-                  single:
-                    upstream:
-                      name: default-example-svc-80
-                      namespace: default
-    bindAddress: '::'
-    bindPort: 9080
-    name: http
-metadata:
-  labels:
-    created_by: gloo-kube-gateway-api-translator
-  name: example-gateway
-  namespace: default
